Made three factory functions that return methods
Gameboard factory function contain an event listener for each cells
this click event will retrieve the current player's mark
and drop it on the cell that was clicked
it will display the current player's turn in the statusDiv
and will correctly switch to the other player after every turn
because it can access both displayturn and playgame methods
from ScreenDisplay and GameFlow .

GameFlow factory function will control the flow of the game
it has an array called players, which contains two dict
these dict store info about the name and the mark of the 2 players
Gameflow has activePlayer variable that stores an element of the array
then getPlayer method  will retrieve that element
and switchPlayer method will switch between the 2 players
the playGame method calls switchPlayer and for now
also displays the player who is playing in the console thanks
to getActivePlayer

ScreenDisplay factory function will display msg in statusDiv
in order to access switchPlayer and getActivePlayer methods
i made an obj of Gameflow.

reinitiliazeGame is function that when called will re-start the game

I made a couple global variables
i need to be able to access the gridCells the statusDiv
and potentially the board too from within anywhere.
whereas if i declare these variable inside of the Gameboard
then i can only access them directly from within that function

the last line is a call to Gameboard()
you only need to call this function in the global environment to play